+++ runtime/src/main/java/io/quarkiverse/embedded/postgresql/EmbeddedPostgreSQLConfigSource.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.quarkiverse.embedded.postgresql;
+
+import java.util.Map;
+
+import io.smallrye.config.common.MapBackedConfigSource;
+
+public class EmbeddedPostgreSQLConfigSource extends MapBackedConfigSource {
+
+    // this is higher than the file system or jar ordinals, but lower than env vars
+    // https://github.com/quarkusio/quarkus/pull/14777
+    private static final int ORDINAL = 270;
+
+    public EmbeddedPostgreSQLConfigSource(Map<String, String> propertyMap) {
+        super("inmemory-postgresql", propertyMap, ORDINAL, false);
+    }
+}
